+++ solvers/bd_matrix.h
@@ -14,12 +14,12 @@
 typedef struct bd_matrix_t bd_matrix_t;
 
 // This returns a block diagonal matrix with a fixed block size.
-bd_matrix_t* bd_matrix_new(int num_block_rows,
-                           int block_size);
+bd_matrix_t* bd_matrix_new(size_t num_block_rows,
+                           size_t block_size);
 
 // This returns a block diagonal matrix with variable block sizes.
-bd_matrix_t* var_bd_matrix_new(int num_block_rows,
-                               int* block_sizes);
+bd_matrix_t* var_bd_matrix_new(size_t num_block_rows,
+                               size_t* block_sizes);
 
 // Creates a clone of the given matrix.
 bd_matrix_t* bd_matrix_clone(bd_matrix_t* matrix);
@@ -28,10 +28,10 @@ bd_matrix_t* bd_matrix_clone(bd_matrix_t* matrix);
 void bd_matrix_free(bd_matrix_t* matrix);
 
 // Returns the number of block rows in the matrix.
-int bd_matrix_num_block_rows(bd_matrix_t* matrix);
+size_t bd_matrix_num_block_rows(bd_matrix_t* matrix);
 
 // Returns the number of (individual) rows in the matrix.
-int bd_matrix_num_rows(bd_matrix_t* matrix);
+size_t bd_matrix_num_rows(bd_matrix_t* matrix);
 
 // Sets all values in the matrix to zero.
 void bd_matrix_zero(bd_matrix_t* matrix);
@@ -40,7 +40,7 @@ void bd_matrix_zero(bd_matrix_t* matrix);
 void bd_matrix_add_identity(bd_matrix_t* matrix, real_t scale_factor);
 
 // Returns the size of the given block row.
-int bd_matrix_block_size(bd_matrix_t* matrix, int block_row);
+size_t bd_matrix_block_size(bd_matrix_t* matrix, int block_row);
 
 // Inserts the given block into the given block row in the matrix.
 void bd_matrix_insert_block(bd_matrix_t* matrix, int block_row, real_t* block);
